@@ -7,68 +7,69 @@ version: 1.0.10+12
 
 
 environment:
-  sdk: '>=3.0.1 <4.0.0'
+  sdk: '>=3.1.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
 
-  cupertino_icons: ^1.0.5 
+  cupertino_icons: ^1.0.6 
   intl: ^0.18.1
-  uuid: ^3.0.7
+  uuid: ^4.2.1
   collection: ^1.17.1
   dartz: ^0.10.1
-  get_it: 7.6.0
+  get_it: 7.6.4
   # Интерфейсные фичи
-  flutter_hooks: ^0.18.6
+  flutter_hooks: ^0.20.3
   flutter_colorpicker: ^1.0.3
-  url_launcher: ^6.1.11
-  flutter_slidable: ^3.0.0
-  cached_network_image: ^3.2.3
+  url_launcher: ^6.2.1
+  flutter_slidable: ^3.0.1
+  cached_network_image: ^3.3.0
   # Управление состояниями
-  injectable: ^2.1.2 
+  injectable: ^2.3.2 
   flutter_bloc: ^8.1.3
   equatable: ^2.0.5
-  freezed: ^2.3.5
-  json_serializable: ^6.7.0
+  freezed: ^2.4.5
+  json_serializable: ^6.7.1
   json_annotation: ^4.8.1
-  freezed_annotation: ^2.2.0
+  freezed_annotation: ^2.4.1
   # Для функции поделиться данными тренировки
   screenshot: ^2.1.0
-  share_plus: ^7.0.2
+  share_plus: ^7.2.1
   # Локальное хранение данных в NoSQL базе
   hive: ^2.2.3
   hive_flutter: ^1.1.0
   # Локальное хранение данных в SQL базе
-  drift: ^2.9.0
-  sqlite3_flutter_libs: ^0.5.15
-  path_provider: ^2.0.15
+  drift: ^2.13.1
+  sqlite3_flutter_libs: ^0.5.18
+  path_provider: ^2.1.1
   path: ^1.8.3
   #Интеграция со службами Firebase 
-  firebase_core: ^2.14.0
-  firebase_analytics: ^10.4.3
-  firebase_crashlytics: ^3.3.3
+  firebase_core: ^2.22.0
+  firebase_analytics: ^10.6.4
+  firebase_crashlytics: ^3.4.4
+  firebase_app_check: ^0.2.1+4
   #Удаленная база данных
-  cloud_firestore: ^4.8.1
+  cloud_firestore: ^4.13.0
   #Проверка интернет соединения
   internet_connection_checker: ^1.0.0+1
   #Аунтификация
-  firebase_auth: ^4.6.3
+  firebase_auth: ^4.13.0
 
   #Выбор изображения с устройства
-  image_picker: ^0.8.9
+  image_picker: ^1.0.4
   
 
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^2.4.5
-  injectable_generator: ^2.1.6
-  drift_dev: ^2.8.3
+  build_runner: ^2.4.6
+  injectable_generator: ^2.4.1
+  drift_dev: ^2.13.2
 
-  flutter_lints: ^2.0.1
+  flutter_lints: ^3.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
