@@ -1,15 +1,20 @@
+import 'package:flutter/material.dart';
 import 'package:my_climbing_trek/features/traveling/domain/entities/transport_type.dart';
+import 'package:my_climbing_trek/features/traveling/domain/entities/travel_day_line.dart';
 
-class TransportLine {
+class TransportLine extends TravelDayLine {
   final TransportType type;
-  final DateTime start;
-  final DateTime? finish;
-  final String description;
+  final TimeOfDay? finishTime;
+
+  @override
+  IconData? get icon => type.icon;
 
   TransportLine({
     required this.type,
-    required this.start,
-    this.finish,
-    required this.description,
+    required super.name,
+    required super.description,
+    required super.time,
+    super.contacts,
+    this.finishTime,
   });
 }
