@@ -17,6 +17,9 @@ final _privateConstructorUsedError = UnsupportedError(
 /// @nodoc
 mixin _$CurrentTravelState {
   Travel? get travel => throw _privateConstructorUsedError;
+  TravelDay? get travelDay => throw _privateConstructorUsedError;
+  bool get useFilter => throw _privateConstructorUsedError;
+  String get errorMessage => throw _privateConstructorUsedError;
 
   @JsonKey(ignore: true)
   $CurrentTravelStateCopyWith<CurrentTravelState> get copyWith =>
@@ -29,7 +32,11 @@ abstract class $CurrentTravelStateCopyWith<$Res> {
           CurrentTravelState value, $Res Function(CurrentTravelState) then) =
       _$CurrentTravelStateCopyWithImpl<$Res, CurrentTravelState>;
   @useResult
-  $Res call({Travel? travel});
+  $Res call(
+      {Travel? travel,
+      TravelDay? travelDay,
+      bool useFilter,
+      String errorMessage});
 }
 
 /// @nodoc
@@ -46,12 +53,27 @@ class _$CurrentTravelStateCopyWithImpl<$Res, $Val extends CurrentTravelState>
   @override
   $Res call({
     Object? travel = freezed,
+    Object? travelDay = freezed,
+    Object? useFilter = null,
+    Object? errorMessage = null,
   }) {
     return _then(_value.copyWith(
       travel: freezed == travel
           ? _value.travel
           : travel // ignore: cast_nullable_to_non_nullable
               as Travel?,
+      travelDay: freezed == travelDay
+          ? _value.travelDay
+          : travelDay // ignore: cast_nullable_to_non_nullable
+              as TravelDay?,
+      useFilter: null == useFilter
+          ? _value.useFilter
+          : useFilter // ignore: cast_nullable_to_non_nullable
+              as bool,
+      errorMessage: null == errorMessage
+          ? _value.errorMessage
+          : errorMessage // ignore: cast_nullable_to_non_nullable
+              as String,
     ) as $Val);
   }
 }
@@ -64,7 +86,11 @@ abstract class _$$_CurrentTravelStateCopyWith<$Res>
       __$$_CurrentTravelStateCopyWithImpl<$Res>;
   @override
   @useResult
-  $Res call({Travel? travel});
+  $Res call(
+      {Travel? travel,
+      TravelDay? travelDay,
+      bool useFilter,
+      String errorMessage});
 }
 
 /// @nodoc
@@ -79,12 +105,27 @@ class __$$_CurrentTravelStateCopyWithImpl<$Res>
   @override
   $Res call({
     Object? travel = freezed,
+    Object? travelDay = freezed,
+    Object? useFilter = null,
+    Object? errorMessage = null,
   }) {
     return _then(_$_CurrentTravelState(
       travel: freezed == travel
           ? _value.travel
           : travel // ignore: cast_nullable_to_non_nullable
               as Travel?,
+      travelDay: freezed == travelDay
+          ? _value.travelDay
+          : travelDay // ignore: cast_nullable_to_non_nullable
+              as TravelDay?,
+      useFilter: null == useFilter
+          ? _value.useFilter
+          : useFilter // ignore: cast_nullable_to_non_nullable
+              as bool,
+      errorMessage: null == errorMessage
+          ? _value.errorMessage
+          : errorMessage // ignore: cast_nullable_to_non_nullable
+              as String,
     ));
   }
 }
@@ -92,14 +133,24 @@ class __$$_CurrentTravelStateCopyWithImpl<$Res>
 /// @nodoc
 
 class _$_CurrentTravelState implements _CurrentTravelState {
-  const _$_CurrentTravelState({this.travel});
+  const _$_CurrentTravelState(
+      {this.travel,
+      this.travelDay,
+      required this.useFilter,
+      required this.errorMessage});
 
   @override
   final Travel? travel;
+  @override
+  final TravelDay? travelDay;
+  @override
+  final bool useFilter;
+  @override
+  final String errorMessage;
 
   @override
   String toString() {
-    return 'CurrentTravelState(travel: $travel)';
+    return 'CurrentTravelState(travel: $travel, travelDay: $travelDay, useFilter: $useFilter, errorMessage: $errorMessage)';
   }
 
   @override
@@ -107,11 +158,18 @@ class _$_CurrentTravelState implements _CurrentTravelState {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$_CurrentTravelState &&
-            (identical(other.travel, travel) || other.travel == travel));
+            (identical(other.travel, travel) || other.travel == travel) &&
+            (identical(other.travelDay, travelDay) ||
+                other.travelDay == travelDay) &&
+            (identical(other.useFilter, useFilter) ||
+                other.useFilter == useFilter) &&
+            (identical(other.errorMessage, errorMessage) ||
+                other.errorMessage == errorMessage));
   }
 
   @override
-  int get hashCode => Object.hash(runtimeType, travel);
+  int get hashCode =>
+      Object.hash(runtimeType, travel, travelDay, useFilter, errorMessage);
 
   @JsonKey(ignore: true)
   @override
@@ -122,11 +180,20 @@ class _$_CurrentTravelState implements _CurrentTravelState {
 }
 
 abstract class _CurrentTravelState implements CurrentTravelState {
-  const factory _CurrentTravelState({final Travel? travel}) =
-      _$_CurrentTravelState;
+  const factory _CurrentTravelState(
+      {final Travel? travel,
+      final TravelDay? travelDay,
+      required final bool useFilter,
+      required final String errorMessage}) = _$_CurrentTravelState;
 
   @override
   Travel? get travel;
+  @override
+  TravelDay? get travelDay;
+  @override
+  bool get useFilter;
+  @override
+  String get errorMessage;
   @override
   @JsonKey(ignore: true)
   _$$_CurrentTravelStateCopyWith<_$_CurrentTravelState> get copyWith =>
